Add relations to Result domain and relation model

+++ src/main/resources/ordinacija_ddl/app_user.sql
@@ -4,7 +4,7 @@ SET FOREIGN_KEY_CHECKS=1;
 
 CREATE TABLE  `app_user` (
 
-  `id`             INT(11) NOT NULL AUTO_INCREMENT, 
+  `id`             BIGINT(20) NOT NULL AUTO_INCREMENT, 
   `username`       VARCHAR(50) NOT NULL,
   `password`       VARCHAR(255) DEFAULT NULL,
   `first_name`     VARCHAR(50) DEFAULT NULL,
