+++ src/main/resources/ordinacija_ddl/medical.sql
@@ -1,7 +1,7 @@
 CREATE TABLE `medical` (
 
   `id`            BIGINT(20) NOT NULL AUTO_INCREMENT,
-  `doctor_id`     INT(11) DEFAULT NULL,
+  `doctor_id`     BIGINT(20) DEFAULT NULL,
   `patient_id`    BIGINT(20) DEFAULT NULL,
   `medical_date`  DATETIME DEFAULT NULL,
   `description`   TEXT,
